fix: unbreak the never-compiled unit tests in three bin crates, and gate them
`display_server`, `usb_hid` and `usbhub` are `no_std` binaries whose
`#[cfg(test)]` modules had never compiled, so those tests had never run
once. Each failed differently:

- `display_server` declared `#![no_std]`/`#![no_main]` and its allocator,
  `#[alloc_error_handler]` and `#[panic_handler]` unconditionally, so the
  std test harness hit duplicate lang items.
- `usb_hid` the same.
- `usbhub`'s crate-level gating was already correct. Its actual bug was a
  `#[cfg(not(test))]` and the "Hub daemon main" doc comment having drifted
  upward onto `enumerate_hubs_once` — which already had its own — leaving
  `fn program_main` ungated. Under `cfg(test)` the whole daemon body then
  compiled and referenced syscall-only symbols, producing all 28 errors.

All three now use the `cfg_attr(not(test), no_std)` pattern the tree
already applies in `term`, `xhci`, `e1000`, `igb` and others: under
`cfg(test)` std supplies the lang items and the conflict disappears.
Gating is kept as narrow as possible — over-gating `use` statements is
what made `usbhub` look like an import problem.

**54 previously-written tests now execute for the first time** (48 in
`display_server` — animation, config, decoration, keybind, workspace; 6 in
`usbhub`) and all pass, so nothing had rotted. `usb_hid` turned out to
have no tests at all rather than dead ones; it gains 21 covering
`classify_role` and `hid_report_descriptor_len`, the two genuinely pure
helpers — the latter verified to fail closed on truncated, zero-length and
over-claiming descriptors rather than reading past its buffer.

All three are wired into `USERSPACE_LIB_HOST_TEST_PACKAGES` with no extra
args. That is deliberate and load-bearing: they are bin-only crates with
no `[lib]` target, so the `--lib` several neighbouring entries carry would
have matched nothing and skipped every test while still reporting green.
Proven by a negative control — breaking one assertion fails
`cargo xtask check` with exit 1, and cargo's own rerun hint names
`--bin display_server`.

Two follow-on cleanups the newly-compiling tests made possible:

- Four `#[allow(dead_code)]` in `display_server` were proven unnecessary
  (determined empirically — every allow stripped, the production lint lane
  re-run, only the 37 the compiler still demanded put back) along with a
  doc paragraph claiming `LABEL_CTL_CMD` was unconsumed. It has callers.
- `usbhub`'s `hub_backoff_grows_after_threshold` asserted `grown >= base`,
  which a curve flattened to a constant would satisfy — precisely the
  regression the test is named for. Tightened to `>`.

`display_server` carries a new crate-level `#![cfg_attr(test, allow(dead_code))]`:
under `cfg(test)` nothing calls `program_main`, so all ~11.5k lines read as
dead. The alternative was `#[cfg(not(test))]` on ~121 items across 13
files, which risks changing production behaviour to satisfy a lint. The
production lane keeps `dead_code` fully enforced and is clippy-clean, so
nothing dead can reach the shipped binary unnoticed.

Gates: check, smoke-test, regression 11/11, tiling, clipboard, screenshot,
term-daily-driver (6 arms), usb-smoke, usb-hub-smoke, usb-report-smoke.

+++ userspace/display_server/src/keybind.rs
@@ -381,7 +381,7 @@ mod tests {
         {
             let bid = table
                 .table
-                .match_bind(MOD_SUPER, kc.0 as u32)
+                .match_bind(MOD_SUPER, kc.0)
                 .expect("super+digit bound");
             assert_eq!(
                 table.action(bid),
@@ -394,11 +394,11 @@ mod tests {
     fn push_resize_mode_overlays_table() {
         let mut stack = BindStack::new();
         // No `H` binding in default mode.
-        assert!(stack.active_table().match_bind(0, KEY_H.0 as u32).is_none());
+        assert!(stack.active_table().match_bind(0, KEY_H.0).is_none());
         stack.push_mode(BindMode::Resize);
         let bid = stack
             .active_table()
-            .match_bind(0, KEY_H.0 as u32)
+            .match_bind(0, KEY_H.0)
             .expect("H bound in resize mode");
         assert!(matches!(
             stack.lookup_action(bid),
@@ -408,7 +408,7 @@ mod tests {
             })
         ));
         stack.pop_mode();
-        assert!(stack.active_table().match_bind(0, KEY_H.0 as u32).is_none());
+        assert!(stack.active_table().match_bind(0, KEY_H.0).is_none());
     }
 
     #[test]
@@ -428,7 +428,7 @@ mod tests {
         stack.push_mode(BindMode::Resize);
         let bid = stack
             .active_table()
-            .match_bind(0, KEY_H.0 as u32)
+            .match_bind(0, KEY_H.0)
             .expect("H bound in resize mode");
         match stack.lookup_action(bid) {
             Some(KeybindAction::ResizeFocused { step, .. }) => {
